# Random: always use SamplerRangeFast for MersenneTwister