@StefanKarpinski 2018-06-13
We can break some things between `0.7.0-alpha` and `0.7.0-beta` so this would be ok.