@rfourquet 2018-06-14
Thanks for your answers! So I will mark this for triage. 

> I am still fascinated that a single division is so expensive compared to our MersenneTwister entropy generation

Me too, indeed! SIMD is certainly crucial for that...

If someone likes to test the performance on her machine, it should be rather straigthforward, with this `eval` rather than compiling this branch:
```julia
using Random
for T in Base.BitInteger_types
    @eval Random Sampler(rng::MersenneTwister, r::UnitRange{$T}, ::Val{Inf}) = SamplerRangeFast(r)
end
```