@rfourquet 2018-06-22
Thanks for having discussed this.

I did few more benchmark to try to get an idea what would be the expected speed-up for a range of "random length" (not so trivial to measure that in a meaningful way; I did what I could). Benchmarking only the step 2) mentioned in the OP (which is equivalent to benchmarking `rand!` on asymptotically large arrays, which favors `SamplerRangeInt`), I observe speed-ups typically between 40% and 60% (and a fair majority over 50%), so it seems safe to assume that it will be an improvement on most machines. The main drawback I see is that the time it takes is less predictable (it depends a lot on the length of the range).

If no objection comes, I will therefore merge this week-end.