@mschauer 2018-06-22
The improvement even on asymptotically large arrays settles it.