@rfourquet 2018-06-25
The Travis [failure](https://travis-ci.org/JuliaLang/julia/jobs/396101554) is irrelated to this change and happens in other PRs too (concerns `FileWatching`).